:sparkles:Add to/from dict/dataframe methods to Material
